requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
psycopg2-binary==2.9.11
python-multipart==0.0.20
sqlalchemy
pydantic
python-dotenv
pytz
skyfield
requests
openai
tiktoken
aiofiles
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyswisseph
geopy
pytest
pytest-asyncio
alembic